fixed rating model

+++ lib/models/rating.dart
@@ -5,6 +5,6 @@ class Rating {
   Rating({required this.rate, required this.count});
 
   factory Rating.fromJson(Map<String, dynamic> json) {
-    return Rating(rate: json['rate'], count: json['count']);
+    return Rating(rate: (json['rate']).toDouble(), count: json['count']);
   }
 }
